Add CardinalAdapter benchmarks for memory and speed tests
- Implemented memory and speed benchmarks for Cardinal peak picking, baseline reduction, noise reduction, and normalization in respective test files.
- Updated test fixtures to utilize a new example data file for consistency across tests.
- Adjusted batch sizes in flat generators to optimize performance.
- Enhanced logging for benchmark tests to provide clearer insights during execution.
- Updated documentation to include new Cardinal benchmarks in the testing framework.

+++ src/massflow/preprocess/helper/peak_align_helper.py
@@ -421,14 +421,14 @@ def peak_aligner(
     reference: NDArray[np.float64],
     tolerance: float,
     units: str = "ppm",
-) -> NDArray[np.float64]:
+) -> NDArray[np.float32]:
     """Align a flat batch of spectra to a reference m/z axis."""
     norm_units = _normalize_units(units)
     tol_method = "x" if norm_units == "relative" else "abs"
     code = get_method_code(tol_method)
 
     mz_arr = np.asarray(mz_data, dtype=np.float64)
-    intensity_arr = np.asarray(intensity, dtype=np.float64)
+    intensity_arr = np.asarray(intensity, dtype=np.float32)
     lengths_arr = np.asarray(lengths, dtype=np.int32)
     reference_arr = np.asarray(reference, dtype=np.float64)
